transform script em pacote python
