+++ makefile
@@ -4,7 +4,11 @@ download:
 	ckanapi dump datasets --remote https://dados.mg.gov.br/ --datapackages=datasets despesa
 
 load: 
-	python main.py
+	python -m dadosmg.load
 
-all: load
+transform:
+	python -m dadosmg.transform ft_despesa
+	python -m dadosmg.transform dm_empenho_desp
+
+all: load transform
