+++ src/dadosmg/__init__.py
@@ -1,0 +1,6 @@
+import logging
+
+LOG_FORMAT = '%(asctime)s %(levelname)-5.5s [%(name)s] %(message)s'
+LOG_DATE_FORMAT = '%Y-%m-%dT%H:%M:%S%z'
+logging.basicConfig(format=LOG_FORMAT, datefmt=LOG_DATE_FORMAT, level=logging.INFO)
+logger = logging.getLogger(__name__)
